requirements.txt
python-dotenv
openai
newsapi-python
requests